+++ packages/moonstone/Scroller/Scrollbar.js
@@ -346,6 +346,7 @@ class ScrollbarBase extends PureComponent {
 			<div ref={this.initContainerRef} className={containerClassName}>
 				<ScrollButton
 					data-scroll-button="previous"
+					data-spotlight-overflow="ignore"
 					direction={vertical ? 'up' : 'left'}
 					disabled={disabled || prevButtonDisabled}
 					onClick={this.handlePrevScroll}
@@ -364,6 +365,7 @@ class ScrollbarBase extends PureComponent {
 				/>
 				<ScrollButton
 					data-scroll-button="next"
+					data-spotlight-overflow="ignore"
 					direction={vertical ? 'down' : 'right'}
 					disabled={disabled || nextButtonDisabled}
 					onClick={this.handleNextScroll}