+++ packages/moonstone/Scroller/ScrollerNative.js
@@ -4,8 +4,6 @@
  * The default export is {@link moonstone/Scroller.ScrollerNative}.
  */
 
-import SpotlightContainerDecorator from '@enact/spotlight/SpotlightContainerDecorator';
-
 import ScrollableNative from './ScrollableNative';
 import {ScrollerBase} from './Scroller';
 
@@ -25,11 +23,8 @@ import {ScrollerBase} from './Scroller';
  * @ui
  * @private
  */
-const ScrollerNative = SpotlightContainerDecorator(
-	{restrict: 'self-first'},
-	ScrollableNative(
-		ScrollerBase
-	)
+const ScrollerNative = ScrollableNative(
+	ScrollerBase
 );
 
 // Docs for ScrollerNative