+++ packages/moonstone/CHANGELOG.md
@@ -5,10 +5,14 @@ The following is a curated list of changes in the Enact moonstone module, newest
 ## [Unreleased]
 
 ### Deprecated
+- `moonstone/TextSizeDecorator` and it will be replaced by `moonstone/AccessibilityDecorator`
 
 ### Added
 
 - `moonstone/TooltipDecorator` config property to direct tooltip into a property instead of adding to `children`
+- `moonstone/VideoPlayer` prop `thumbnailUnavailable` to fade thumbnail
+- `moonstone/AccessibilityDecorator` with `highContrast` and `textSize`
+- `moonstone/VideoPlayer` high contrast scrim
 
 ### Changed
 
@@ -28,8 +32,15 @@ The following is a curated list of changes in the Enact moonstone module, newest
 - `moonstone/FormCheckbox` to display correct colors in light skin
 - `moonstone/Picker` and `moonstone/RangePicker` to forward `onKeyDown` events when not `joined`
 - `moonstone/SelectableItem` to display correct icon width and alignment
+- `moonstone/LabeledItem` to always match alignment with the locale
 - `moonstone/Scroller` to properly 5-way navigate from scroll buttons
+- `moonstone/Scroller` and `moonstone/VirtualList` navigation via 5-way from paging controls
 - `moonstone/ExpandableList` to display correct font weight and size for list items
+- `moonstone/Divider` to not italicize in non-italic locales
+
+### Removed
+
+- `moonstone/Input` large-text mode
 
 ## [1.6.1] - 2017-08-07